+++ .gitignore
@@ -1,0 +1,6 @@
+*.env
+__pycache__/
+db.sqlite3
+*.pyc
+*.log
+migrations/
