+++ PedoController/api/views/__init__.py
@@ -1,0 +1,8 @@
+from .Api import *
+from .Baiter import *
+from .Conversation import *
+from .Message import *
+from .Pedophile import *
+from .SocialNetwork import *
+
+#__all__ = ["Api", "Baiter", "Conversation", "Message", "Pedophile", "SocialNetwork"]
